+++ ts/components/icons.ts
@@ -5,9 +5,11 @@
 
 export { default as hsplitIcon } from "@mdi/svg/svg/arrow-split-horizontal.svg";
 export { default as vsplitIcon } from "@mdi/svg/svg/arrow-split-vertical.svg";
+export { default as manualIcon } from "@mdi/svg/svg/book-open-variant.svg";
 export { default as chevronDown } from "@mdi/svg/svg/chevron-down.svg";
 export { default as chevronLeft } from "@mdi/svg/svg/chevron-left.svg";
 export { default as chevronRight } from "@mdi/svg/svg/chevron-right.svg";
 export { default as chevronUp } from "@mdi/svg/svg/chevron-up.svg";
 export { default as horizontalHandle } from "@mdi/svg/svg/drag-horizontal.svg";
 export { default as verticalHandle } from "@mdi/svg/svg/drag-vertical.svg";
+export { default as infoCircle } from "bootstrap-icons/icons/info-circle.svg";
