+++ ts/components/types.ts
@@ -3,3 +3,9 @@
 
 export type Size = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
 export type Breakpoint = "xs" | "sm" | "md" | "lg" | "xl" | "xxl";
+
+export type HelpItem = {
+    title: string;
+    help?: string;
+    url?: string;
+};
