+++ ts/deck-options/LapseOptions.svelte
@@ -9,15 +9,15 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
     import type Modal from "bootstrap/js/dist/modal";
 
     import DynamicallySlottable from "../components/DynamicallySlottable.svelte";
+    import HelpModal from "../components/HelpModal.svelte";
     import Item from "../components/Item.svelte";
     import TitledContainer from "../components/TitledContainer.svelte";
+    import type { HelpItem } from "../components/types";
     import EnumSelectorRow from "./EnumSelectorRow.svelte";
-    import HelpModal from "./HelpModal.svelte";
     import type { DeckOptionsState } from "./lib";
     import SettingTitle from "./SettingTitle.svelte";
     import SpinBoxRow from "./SpinBoxRow.svelte";
     import StepsInputRow from "./StepsInputRow.svelte";
-    import type { DeckOption } from "./types";
     import Warning from "./Warning.svelte";
 
     export let state: DeckOptionsState;
@@ -61,7 +61,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
             url: HelpPage.Leeches.waiting,
         },
     };
-    const helpSections = Object.values(settings) as DeckOption[];
+    const helpSections = Object.values(settings) as HelpItem[];
 
     let modal: Modal;
     let carousel: Carousel;
