+++ ts/deck-options/icons.ts
@@ -3,8 +3,6 @@
 
 /// <reference types="../lib/image-import" />
 
-export { default as manualIcon } from "@mdi/svg/svg/book-open-variant.svg";
 export { default as chevronDown } from "@mdi/svg/svg/chevron-down.svg";
 export { default as revertIcon } from "bootstrap-icons/icons/arrow-counterclockwise.svg";
 export { default as gearIcon } from "bootstrap-icons/icons/gear.svg";
-export { default as infoCircle } from "bootstrap-icons/icons/info-circle.svg";
