--- ts/deck-options/types.ts
@@ -1,8 +1,0 @@
-// Copyright: Ankitects Pty Ltd and contributors
-// License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-
-export type DeckOption = {
-    title: string;
-    help?: string;
-    url?: string;
-};
