+++ ts/lib/help-page.ts
@@ -33,4 +33,7 @@ export const HelpPage = {
     Studying: {
         siblingsAndBurying: "https://docs.ankiweb.net/studying.html#siblings-and-burying",
     },
+    Importing: {
+        ankiPackage: "https://docs.ankiweb.net/importing.html#packaged-decks",
+    },
 };
